Fashion_CV/pyproject.toml
[project]
name = "fashion-cv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.6",
    "pillow>=12.2.0",
    "tqdm>=4.67.3",
]
